+++ src/cgexplore/systems_optimisation.py
@@ -5,5 +5,15 @@ from cgexplore._internal.systems_optimisation.inputs import (
     Chromosome,
     ChromosomeGenerator,
 )
+from cgexplore._internal.systems_optimisation.utilities import (
+    get_neighbour_library,
+    yield_near_models,
+)
 
-__all__ = ["ChromosomeGenerator", "Chromosome", "Generation"]
+__all__ = [
+    "ChromosomeGenerator",
+    "Chromosome",
+    "Generation",
+    "yield_near_models",
+    "get_neighbour_library",
+]
